requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
httpx>=0.27.2
pydantic>=2.10.0
python-dotenv>=1.0.1
